+++ android/src/main/kotlin/me/yohom/foundation_fluttify/android/util/PairHandler.kt
@@ -3,7 +3,6 @@ package me.yohom.foundation_fluttify.android.util
 import android.util.Pair
 import io.flutter.plugin.common.MethodChannel
 import me.yohom.foundation_fluttify.HEAP
-import me.yohom.foundation_fluttify.fluttifySequence
 import me.yohom.foundation_fluttify.jsonable
 
 fun PairHandler(method: String, args: Map<String, Any>, methodResult: MethodChannel.Result) {
@@ -15,9 +14,9 @@ fun PairHandler(method: String, args: Map<String, Any>, methodResult: MethodChan
             if (pair.first.jsonable()) {
                 methodResult.success(pair.first)
             } else {
-                val seqNumber = fluttifySequence
-                HEAP[seqNumber] = pair.first
-                methodResult.success(seqNumber)
+                val hash = System.identityHashCode(pair.first)
+                HEAP[hash] = pair.first
+                methodResult.success(hash)
             }
         }
         "android.util.Pair::getSecond" -> {
@@ -27,9 +26,9 @@ fun PairHandler(method: String, args: Map<String, Any>, methodResult: MethodChan
             if (pair.second.jsonable()) {
                 methodResult.success(pair.second)
             } else {
-                val seqNumber = fluttifySequence
-                HEAP[seqNumber] = pair.second
-                methodResult.success(seqNumber)
+                val hash = System.identityHashCode(pair.second)
+                HEAP[hash] = pair.second
+                methodResult.success(hash)
             }
         }
         else -> methodResult.notImplemented()