+++ ios/Classes/CLHeadingHandler.m
@@ -11,7 +11,6 @@
 extern NSMutableDictionary<NSString *, NSObject *> *STACK;
 extern NSMutableDictionary<NSNumber *, NSObject *> *HEAP;
 extern BOOL enableLog;
-extern int getFluttifySequence(void);
 
 void CLHeadingHandler(NSString* method, NSDictionary* args, FlutterResult methodResult){
     if ([@"CLHeading::getMagneticHeading" isEqualToString:method]) {