+++ ios/Classes/CLLocationCoordinate2DHandler.m
@@ -8,7 +8,6 @@
 extern NSMutableDictionary<NSString *, NSObject *> *STACK;
 extern NSMutableDictionary<NSNumber *, NSObject *> *HEAP;
 extern BOOL enableLog;
-extern int getFluttifySequence(void);
 
 void CLLocationCoordinate2DHandler(NSString* method, NSDictionary* args, FlutterResult methodResult) {
     // CLLocationCoordinate2D获取latitude
@@ -41,10 +40,9 @@ void CLLocationCoordinate2DHandler(NSString* method, NSDictionary* args, Flutter
         CLLocationCoordinate2D data = CLLocationCoordinate2DMake(latitude, longitude);
         
         NSValue *dataValue = [NSValue value:&data withObjCType:@encode(CLLocationCoordinate2D)];
-        int64_t seqNumber = getFluttifySequence();
-        HEAP[@(seqNumber)] = dataValue;
+        HEAP[@(dataValue.hash)] = dataValue;
         
-        methodResult(@(seqNumber));
+        methodResult(@(dataValue.hash));
     }
     // 批量创建CLLocationCoordinate2D
     else if ([@"CLLocationCoordinate2D::create_batchCLLocationCoordinate2D" isEqualToString:method]) {
@@ -59,9 +57,8 @@ void CLLocationCoordinate2DHandler(NSString* method, NSDictionary* args, Flutter
             CLLocationCoordinate2D data = CLLocationCoordinate2DMake(latitude, longitude);
 
             NSValue *dataValue = [NSValue value:&data withObjCType:@encode(CLLocationCoordinate2D)];
-            int64_t seqNumber = getFluttifySequence();
-            [resultBatch addObject:@(seqNumber)];
-            HEAP[@(seqNumber)] = dataValue;
+            [resultBatch addObject:@(dataValue.hash)];
+            HEAP[@(dataValue.hash)] = dataValue;
         }
         
         methodResult(resultBatch);